app/requirements.txt
Django==3.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
graphene-django==2.14.0
django-filter==21.1
Pillow==8.3.2
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-extensions==3.1.5
django_seed==0.3.1
django-phonenumber-field[phonenumbers]